Include backup/restore of UBOOT and recovery to maintain backup consistency
Since MTK6589 BOOT partition is totally dependent on UBOOT we must make
sure that the corresponding UBOOT is backed up/restored together with
BOOT, otherwise device could be soft-bricked after restoration of
previous backup having incompatible boot/uboot pair.

Additionally, we have to make sure also recovery that is compatible
with restored boot/uboot pair is restored so device is able to boot
into recovery after restore.
The drawback of this approach is that in specific cases newer version of recovery
already present in the device will get replaced by the older one.
But... This is still the better case than to end up with
recovery that is totally unbootable...

This change makes backups fully consistent.

Change-Id: Id6ed36ae250b7b4fc54faa5c10ac327f0b623977